refactor: DatabaseProvider 제작
